data-pipelines/esg-gdrive-to-local/requirements.txt
python-dotenv==1.0.1
google-api-python-client==2.157.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1